Call exec from onAfterDownload and move execSync call to function
